Don't build the GTK+ IC itself against dbus
RevBy: Michael Hasselmann
With respect to the IC, DBus is an internal detail
and should not be exposed to it. This separation ensures
that we can switch DBus/IPC implementation without having
to change code in the IC itself.

Full, original commit at 0901b5450be6afa109c1fcc3932d26aa233d4b1f in maliit-framework