Use .pri files for all interdependencies between components
RevBy: TrustMe
This hides the details of how one component needs to use
another component, which is useful when this changes
depending on platform. Also makes it easier to do changes
because the code is in one place and the dependencies
of a given component is clearer.

Full, original commit at c469234b48f33203b51436d46adb31d38600d906 in maliit-framework