Add private MaliitAttributeExtension constructor.
RevBy: Jan Arne Petersen

This constructor is going to be used by MaliitSettingsManager to
create MaliitAttributeExtension with specific id. Also, loosened a bit
strictness of keys and extension ids. with settings introduced it is
possible to have a key in form /a/b/c/d and we were allowing only
/a/b/c. Also now we are getting extension ids lower than zero from
maliit-server, so we have to allow them in property installer.

Full, original commit at 5841838c84badde19983f34388d4eb61de8deaf1 in maliit-framework