Remove singleton pattern for MeegoIMProxy and MeegoIMContextDBusObj
RevBy: NeedsRev
This ruined testability by not allowing to create clean instances,
and was generally misleading as the instances were in practice
managed by MeegoImConnector. Now they are normal members owned by
MeegoImConnector.
MeegoImConnector is still a singleton, so the object lifetimes
are not changed in the typical case. The exception is when
using the private API for MeegoImConnector to manage the instance.
This is what tests will do.

Full, original commit at 1818b8ebe99d4395613d1ee3a80714b9e0259ae9 in maliit-framework